Imp splitting of preprocessing.py
